+++ src/distribution/fisher_snedecor.rs
@@ -49,6 +49,8 @@ impl FisherSnedecor {
     pub fn new(freedom_1: f64, freedom_2: f64) -> Result<FisherSnedecor> {
         if freedom_1.is_nan() || freedom_2.is_nan() {
             Err(StatsError::BadParams)
+        } else if freedom_1 == f64::INFINITY || freedom_2 == f64::INFINITY {
+            Err(StatsError::BadParams)
         } else if freedom_1 <= 0.0 || freedom_2 <= 0.0 {
             Err(StatsError::BadParams)
         } else {
@@ -136,9 +138,9 @@ impl Univariate<f64, f64> for FisherSnedecor {
     /// Calculates the cumulative distribution function for the fisher-snedecor distribution
     /// at `x`
     ///
-    /// # Remarks
+    /// # Panics
     ///
-    /// Returns `NaN` if `freedom_1`, `freedom_2` is `INF`, or `x` is `+INF` or `-INF`
+    /// If `x` is not in `[0, +inf)`
     ///
     /// # Formula
     ///
@@ -150,13 +152,12 @@ impl Univariate<f64, f64> for FisherSnedecor {
     /// the second degree of freedom, and `I` is the regularized incomplete
     /// beta function
     fn cdf(&self, x: f64) -> f64 {
-        if self.freedom_1 == f64::INFINITY || self.freedom_2 == f64::INFINITY || x.is_infinite() {
-            f64::NAN
-        } else {
-            beta::beta_reg(self.freedom_1 / 2.0,
-                           self.freedom_2 / 2.0,
-                           self.freedom_1 * x / (self.freedom_1 * x + self.freedom_2))
-        }
+        assert!(x >= 0.0 && x < f64::INFINITY,
+                format!("{}",
+                        StatsError::ArgIntervalExclMax("x", 0.0, f64::INFINITY)));
+        beta::beta_reg(self.freedom_1 / 2.0,
+                       self.freedom_2 / 2.0,
+                       self.freedom_1 * x / (self.freedom_1 * x + self.freedom_2))
     }
 }
 
@@ -400,26 +401,14 @@ mod test {
         assert_almost_eq!(expected, x, acc);
     }
 
-    fn test_is_nan<F>(freedom_1: f64, freedom_2: f64, eval: F)
-        where F: Fn(FisherSnedecor) -> f64
-    {
-        assert!(get_value(freedom_1, freedom_2, eval).is_nan())
-    }
-
     #[test]
     fn test_create() {
         create_case(0.1, 0.1);
         create_case(1.0, 0.1);
         create_case(10.0, 0.1);
-        create_case(f64::INFINITY, 0.1);
         create_case(0.1, 1.0);
         create_case(1.0, 1.0);
         create_case(10.0, 1.0);
-        create_case(f64::INFINITY, 10.0);
-        create_case(0.1, f64::INFINITY);
-        create_case(1.0, f64::INFINITY);
-        create_case(10.0, f64::INFINITY);
-        create_case(f64::INFINITY, f64::INFINITY);
     }
 
     #[test]
@@ -440,6 +429,9 @@ mod test {
         bad_create_case(0.0, -10.0);
         bad_create_case(-1.0, -10.0);
         bad_create_case(-10.0, -10.0);
+        bad_create_case(f64::INFINITY, 0.1);
+        bad_create_case(0.1, f64::INFINITY);
+        bad_create_case(f64::INFINITY, f64::INFINITY);
     }
 
     #[test]
@@ -453,10 +445,6 @@ mod test {
         test_case(0.1, 10.0, 1.25, |x| x.mean());
         test_case(1.0, 10.0, 1.25, |x| x.mean());
         test_case(10.0, 10.0, 1.25, |x| x.mean());
-        test_is_nan(0.1, f64::INFINITY, |x| x.mean());
-        test_is_nan(1.0, f64::INFINITY, |x| x.mean());
-        test_is_nan(10.0, f64::INFINITY, |x| x.mean());
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.mode());
     }
 
     #[test]
@@ -470,11 +458,6 @@ mod test {
         test_almost(0.1, 10.0, 42.1875, 1e-14, |x| x.variance());
         test_case(1.0, 10.0, 4.6875, |x| x.variance());
         test_case(10.0, 10.0, 0.9375, |x| x.variance());
-        test_is_nan(0.1, f64::INFINITY, |x| x.variance());
-        test_is_nan(1.0, f64::INFINITY, |x| x.variance());
-        test_is_nan(10.0, f64::INFINITY, |x| x.variance());
-        test_is_nan(f64::INFINITY, 10.0, |x| x.variance());
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.mode());
     }
 
     #[test]
@@ -488,11 +471,6 @@ mod test {
         test_almost(0.1, 10.0, 42.1875f64.sqrt(), 1e-14, |x| x.std_dev());
         test_case(1.0, 10.0, 4.6875f64.sqrt(), |x| x.std_dev());
         test_case(10.0, 10.0, 0.9375f64.sqrt(), |x| x.std_dev());
-        test_is_nan(0.1, f64::INFINITY, |x| x.std_dev());
-        test_is_nan(1.0, f64::INFINITY, |x| x.std_dev());
-        test_is_nan(10.0, f64::INFINITY, |x| x.std_dev());
-        test_is_nan(f64::INFINITY, 10.0, |x| x.std_dev());
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.mode());
     }
 
     #[test]
@@ -506,11 +484,6 @@ mod test {
         test_almost(0.1, 10.0, 15.78090735784977089658, 1e-14, |x| x.skewness());
         test_case(1.0, 10.0, 5.773502691896257645091, |x| x.skewness());
         test_case(10.0, 10.0, 3.614784456460255759501, |x| x.skewness());
-        test_is_nan(0.1, f64::INFINITY, |x| x.skewness());
-        test_is_nan(1.0, f64::INFINITY, |x| x.skewness());
-        test_is_nan(10.0, f64::INFINITY, |x| x.skewness());
-        test_is_nan(f64::INFINITY, 10.0, |x| x.skewness());
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.mode());
     }
 
     #[test]
@@ -524,11 +497,6 @@ mod test {
         test_case(10.0, 0.1, 0.0380952380952380952381, |x| x.mode());
         test_case(10.0, 1.0, 4.0 / 15.0, |x| x.mode());
         test_case(10.0, 10.0, 2.0 / 3.0, |x| x.mode());
-        test_is_nan(10.0, f64::INFINITY, |x| x.mode());
-        test_is_nan(f64::INFINITY, 0.1, |x| x.mode());
-        test_is_nan(f64::INFINITY, 1.0, |x| x.mode());
-        test_is_nan(f64::INFINITY, 10.0, |x| x.mode());
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.mode());
     }
 
     #[test]
@@ -542,40 +510,18 @@ mod test {
         test_almost(0.1, 0.1, 0.0234154207226588982471, 1e-16, |x| x.pdf(1.0));
         test_almost(1.0, 0.1, 0.0396064560910663979961, 1e-16, |x| x.pdf(1.0));
         test_almost(10.0, 0.1, 0.0418440630400545297349, 1e-14, |x| x.pdf(1.0));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.pdf(1.0));
         test_almost(0.1, 1.0, 0.0396064560910663979961, 1e-16, |x| x.pdf(1.0));
         test_almost(1.0, 1.0, 0.1591549430918953357689, 1e-16, |x| x.pdf(1.0));
         test_almost(10.0, 1.0, 0.230361989229138647108, 1e-16, |x| x.pdf(1.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.pdf(1.0));
-        test_is_nan(0.1, f64::INFINITY, |x| x.pdf(1.0));
-        test_is_nan(1.0, f64::INFINITY, |x| x.pdf(1.0));
-        test_is_nan(10.0, f64::INFINITY, |x| x.pdf(1.0));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.pdf(1.0));
         test_almost(0.1, 0.1, 0.00221546909694001013517, 1e-18, |x| x.pdf(10.0));
         test_almost(1.0, 0.1, 0.00369960370387922619592, 1e-17, |x| x.pdf(10.0));
         test_almost(10.0, 0.1, 0.00390179721174142927402, 1e-15, |x| x.pdf(10.0));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.pdf(10.0));
         test_almost(0.1, 1.0, 0.00319864073359931548273, 1e-17, |x| x.pdf(10.0));
         test_almost(1.0, 1.0, 0.009150765837179460915678, 1e-17, |x| x.pdf(10.0));
         test_almost(10.0, 1.0, 0.0116493859171442148446, 1e-17, |x| x.pdf(10.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.pdf(10.0));
         test_almost(0.1, 10.0, 0.00305087016058573989694, 1e-15, |x| x.pdf(10.0));
         test_almost(1.0, 10.0, 0.00271897749113479577864, 1e-17, |x| x.pdf(10.0));
         test_almost(10.0, 10.0, 2.4289227234060500084E-4, 1e-18, |x| x.pdf(10.0));
-        test_is_nan(f64::INFINITY, 10.0, |x| x.pdf(10.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.pdf(10.0));
-        test_is_nan(0.1, f64::INFINITY, |x| x.pdf(10.0));
-        test_is_nan(1.0, f64::INFINITY, |x| x.pdf(10.0));
-        test_is_nan(10.0, f64::INFINITY, |x| x.pdf(10.0));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.pdf(10.0));
-        test_is_nan(0.1, 0.1, |x| x.pdf(f64::INFINITY));
-        test_is_nan(0.1, 0.1, |x| x.pdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.pdf(f64::INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.pdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.pdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.pdf(f64::NEG_INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.pdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.pdf(f64::NEG_INFINITY));
     }
 
     #[test]
@@ -583,39 +529,18 @@ mod test {
         test_almost(0.1, 0.1, 0.0234154207226588982471f64.ln(), 1e-15, |x| x.ln_pdf(1.0));
         test_almost(1.0, 0.1, 0.0396064560910663979961f64.ln(), 1e-15, |x| x.ln_pdf(1.0));
         test_almost(10.0, 0.1, 0.0418440630400545297349f64.ln(), 1e-13, |x| x.ln_pdf(1.0));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.ln_pdf(1.0));
         test_almost(0.1, 1.0, 0.0396064560910663979961f64.ln(), 1e-15, |x| x.ln_pdf(1.0));
         test_almost(1.0, 1.0, 0.1591549430918953357689f64.ln(), 1e-15, |x| x.ln_pdf(1.0));
         test_almost(10.0, 1.0, 0.230361989229138647108f64.ln(), 1e-15, |x| x.ln_pdf(1.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.ln_pdf(1.0));
-        test_is_nan(0.1, f64::INFINITY, |x| x.ln_pdf(1.0));
-        test_is_nan(1.0, f64::INFINITY, |x| x.ln_pdf(1.0));
-        test_is_nan(10.0, f64::INFINITY, |x| x.ln_pdf(1.0));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.ln_pdf(1.0));
         test_case(0.1, 0.1, 0.00221546909694001013517f64.ln(), |x| x.ln_pdf(10.0));
         test_almost(1.0, 0.1, 0.00369960370387922619592f64.ln(), 1e-15, |x| x.ln_pdf(10.0));
         test_almost(10.0, 0.1, 0.00390179721174142927402f64.ln(), 1e-13, |x| x.ln_pdf(10.0));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.ln_pdf(10.0));
         test_almost(0.1, 1.0, 0.00319864073359931548273f64.ln(), 1e-15, |x| x.ln_pdf(10.0));
         test_almost(1.0, 1.0, 0.009150765837179460915678f64.ln(), 1e-15, |x| x.ln_pdf(10.0));
         test_case(10.0, 1.0, 0.0116493859171442148446f64.ln(), |x| x.ln_pdf(10.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.ln_pdf(10.0));
         test_almost(0.1, 10.0, 0.00305087016058573989694f64.ln(), 1e-13, |x| x.ln_pdf(10.0));
         test_case(1.0, 10.0, 0.00271897749113479577864f64.ln(), |x| x.ln_pdf(10.0));
         test_almost(10.0, 10.0, 2.4289227234060500084E-4f64.ln(), 1e-14, |x| x.ln_pdf(10.0));
-        test_is_nan(f64::INFINITY, 10.0, |x| x.ln_pdf(10.0));
-        test_is_nan(0.1, f64::INFINITY, |x| x.ln_pdf(10.0));
-        test_is_nan(1.0, f64::INFINITY, |x| x.ln_pdf(10.0));
-        test_is_nan(10.0, f64::INFINITY, |x| x.ln_pdf(10.0));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.ln_pdf(10.0));
-        test_is_nan(0.1, 0.1, |x| x.ln_pdf(f64::INFINITY));
-        test_is_nan(0.1, 0.1, |x| x.ln_pdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.ln_pdf(f64::INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.ln_pdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.ln_pdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.ln_pdf(f64::NEG_INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.ln_pdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.ln_pdf(f64::NEG_INFINITY));
     }
 
     #[test]
@@ -623,34 +548,26 @@ mod test {
         test_almost(0.1, 0.1, 0.44712986033425140335, 1e-15, |x| x.cdf(0.1));
         test_almost(1.0, 0.1, 0.08156522095104674015, 1e-15, |x| x.cdf(0.1));
         test_almost(10.0, 0.1, 0.033184005716276536322, 1e-13, |x| x.cdf(0.1));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.cdf(0.1));
         test_almost(0.1, 1.0, 0.74378710917986379989, 1e-15, |x| x.cdf(0.1));
         test_almost(1.0, 1.0, 0.1949822290421366451595, 1e-16, |x| x.cdf(0.1));
         test_almost(10.0, 1.0, 0.0101195597354337146205, 1e-17, |x| x.cdf(0.1));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.cdf(0.1));
-        test_is_nan(0.1, f64::INFINITY, |x| x.cdf(0.1));
-        test_is_nan(1.0, f64::INFINITY, |x| x.cdf(0.1));
-        test_is_nan(10.0, f64::INFINITY, |x| x.cdf(0.1));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.cdf(0.1));
         test_almost(0.1, 0.1, 0.5, 1e-15, |x| x.cdf(1.0));
         test_almost(1.0, 0.1, 0.16734351500944271141, 1e-14, |x| x.cdf(1.0));
         test_almost(10.0, 0.1, 0.12207560664741704938, 1e-13, |x| x.cdf(1.0));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.cdf(1.0));
         test_almost(0.1, 1.0, 0.83265648499055728859, 1e-15, |x| x.cdf(1.0));
         test_almost(1.0, 1.0, 0.5, 1e-15, |x| x.cdf(1.0));
         test_almost(10.0, 1.0, 0.340893132302059872675, 1e-15, |x| x.cdf(1.0));
-        test_is_nan(f64::INFINITY, 1.0, |x| x.cdf(1.0));
-        test_is_nan(0.1, f64::INFINITY, |x| x.cdf(1.0));
-        test_is_nan(1.0, f64::INFINITY, |x| x.cdf(1.0));
-        test_is_nan(10.0, f64::INFINITY, |x| x.cdf(1.0));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.cdf(1.0));
-        test_is_nan(0.1, 0.1, |x| x.cdf(f64::INFINITY));
-        test_is_nan(0.1, 0.1, |x| x.cdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.cdf(f64::INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.cdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.cdf(f64::INFINITY));
-        test_is_nan(f64::INFINITY, 0.1, |x| x.cdf(f64::NEG_INFINITY));
-        test_is_nan(0.1, f64::INFINITY, |x| x.cdf(f64::NEG_INFINITY));
-        test_is_nan(f64::INFINITY, f64::INFINITY, |x| x.cdf(f64::NEG_INFINITY));
+    }
+
+    #[test]
+    #[should_panic]
+    fn test_cdf_lower_bound() {
+        get_value(0.1, 0.1, |x| x.cdf(-1.0));
+    }
+
+    #[test]
+    #[should_panic]
+    fn test_cdf_upper_bound() {
+        get_value(0.1, 0.1, |x| x.cdf(f64::INFINITY));
     }
 }